requirements.txt
beautifulsoup4
requests
time
Flask
line-bot-sdk
random
urllib.request